Add missing change for abcc: No check paramter code.
We don't really check code for onTransact now. What we really need is the
infomation from the data paramter. Also, this can help loose coupling from
framework since we don't know which the actual parameter code it will be.

This CL should be committed w/ the first edition, this missing is my fault.

Change-Id: I126bb8e290aa509530c87478799d4db4c7eefe0c